refactor: Move CircularTrajectory to its own file
